requirements.txt
ollama>=0.6.1
numpy>=2.3.5

torch>=2.9.1
torchaudio>=2.9.1

onnxruntime>=1.3.0

faster-whisper>=1.2.1
piper-tts>=1.3.0

fastapi[standard]